Rendering keyboard
Implemented the rendering of the keyboard and the styling for it based on the Figma document. For now, the buttons print its letter when pressed. Note: setting handlePress() as the onClick function directly instead of using an arrow function creates some bad behavior where the functions are called immediately when the button is rendered and never on press.