Accessibility additions and lives refactor
Fixed some accessibility issues by making it so the keyboard buttons say what they are when a player is tabbing through the buttons and by having the screen reader provide game updates. Also changed some of the derived values a bit to now keep track of the remaining number of lives and use that to see if the game is lost.